Add augmented and recursive jacobi generalizations
For weight functions with more than one polynomial factor we design
the Augmented Jacobi Orthogonal Polynomials.  Generalized Jacobi can
only handle a single additional factor (with corresponding parameter).
We here add implementations for general products of polynomial factors,
    w(z) = (1-z)**a * (1+z)**b * p1(z)**c1 * ... * pn(z)**cn
with parameters (a, b, c1, ..., cn).

The embedding operators are simple extensions to the single-factor
generalized OPs.  We still need to implement (but first uncover)
the differential structure of this class of OPs.

The recursive_jacobi implementation recursively builds up the
orthogonal polynomials one factor at a time.  This makes indexing
into the parameters a bit unwieldy and the implementation may be
removed in the future.  We prefer the flattened implementation
provided in augmented_jacobi.