Implement Geometry object, fix differential ops for root_h height
Geometry object specifies the stretched cylindrical geometry for a
given problem.  It stores the 'full' or 'half' cylinder_type, the
height polynomial coefficients, the cylinder radius and whether
we are using a root-height domain.  This greatly simplifies calling
the various operators and prevents bugs from forgetting to pass
geometry arguments into functions.

Fixed a bug in the root_h differential operators that was converting
the C parameter too many times.